+++ src/Card.tsx
@@ -2,14 +2,14 @@ import React, { useContext } from 'react'
 import classNames from 'classnames'
 import { ThemeContext } from './context/ThemeContext'
 
-interface Props extends React.HTMLAttributes<HTMLDivElement> {
+export interface CardProps extends React.HTMLAttributes<HTMLDivElement> {
   /**
    * Removes default styles (if true) so you can override with your own background styles
    */
   colored?: boolean
 }
 
-const Card = React.forwardRef<HTMLDivElement, Props>(function Card(props, ref) {
+const Card = React.forwardRef<HTMLDivElement, CardProps>(function Card(props, ref) {
   const { className, children, colored = false, ...other } = props
   const {
     theme: { card },