+++ src/__tests__/Input.test.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { mount } from 'enzyme'
 import Input from '../Input'
+import theme from '../themes/default'
 
 describe('Input', () => {
   it('should render without crashing', () => {
@@ -8,16 +9,14 @@ describe('Input', () => {
   })
 
   it('should render with base styles', () => {
-    const expected =
-      'block w-full text-sm focus:outline-none dark:text-gray-300 form-input leading-5'
+    const expected = theme.input.base
     const wrapper = mount(<Input />)
 
     expect(wrapper.find('input').getDOMNode().getAttribute('class')).toContain(expected)
   })
 
   it('should render with active styles', () => {
-    const expected =
-      'focus:border-purple-400 dark:border-gray-600 focus:shadow-outline-purple dark:focus:border-gray-600 dark:focus:shadow-outline-gray dark:bg-gray-700'
+    const expected = theme.input.active
     const wrapper = mount(<Input />)
 
     expect(wrapper.find('input').getDOMNode().getAttribute('class')).toContain(expected)
@@ -31,24 +30,21 @@ describe('Input', () => {
   })
 
   it('should render with valid styles', () => {
-    const expected =
-      'border-green-600 dark:bg-gray-700 focus:border-green-400 dark:focus:border-green-400 focus:shadow-outline-green dark:focus:shadow-outline-green'
+    const expected = theme.input.valid
     const wrapper = mount(<Input valid />)
 
     expect(wrapper.find('input').getDOMNode().getAttribute('class')).toContain(expected)
   })
 
   it('should render with invalid styles', () => {
-    const expected =
-      'border-red-600 dark:bg-gray-700 focus:border-red-400 dark:focus:border-red-400 focus:shadow-outline-red dark:focus:shadow-outline-red'
+    const expected = theme.input.invalid
     const wrapper = mount(<Input valid={false} />)
 
     expect(wrapper.find('input').getDOMNode().getAttribute('class')).toContain(expected)
   })
 
   it('should render with radio styles', () => {
-    const expected =
-      'text-purple-600 form-radio focus:border-purple-400 focus:outline-none focus:shadow-outline-purple dark:focus:shadow-outline-gray'
+    const expected = theme.input.radio
     const wrapper = mount(<Input type="radio" />)
 
     expect(wrapper.find('input[type="radio"]').getDOMNode().getAttribute('class')).toContain(
@@ -57,8 +53,7 @@ describe('Input', () => {
   })
 
   it('should render with checkbox styles', () => {
-    const expected =
-      'text-purple-600 form-checkbox focus:border-purple-400 focus:outline-none focus:shadow-outline-purple dark:focus:shadow-outline-gray'
+    const expected = theme.input.checkbox
     const wrapper = mount(<Input type="checkbox" />)
 
     expect(wrapper.find('input[type="checkbox"]').getDOMNode().getAttribute('class')).toContain(